chore(DIPRefactored, DIPViolated, DIPViolated2) move package
